Test refactoring (combine 2 test classes)
